backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
google-generativeai>=0.8.3
python-multipart>=0.0.9
websockets>=12.0
python-dotenv>=1.0.0
pydantic>=2.6.0
